requirements.txt
albumentations==1.1.0
python-dotenv==0.19.1
hydra-core==1.1.1
hydra-colorlog==1.1.0
hydra-optuna-sweeper==1.1.1
numpy==1.21.3
optuna==2.10.0
pandas==1.3.4
pytorch_lightning==1.4.9
pytest==6.2.5
PyYAML==6.0
SciencePlots==1.0.9
seaborn==0.11.2
torch==1.10.0
torchvision==0.10.0
wandb==0.12.6
